src/bot/bot_handler.py
from typing import Dict, Any
from slack_bolt import App as SlackApp
from microsoftteams import connectors
from google.cloud import pubsub_v1
from ..config.config_handler import ConfigHandler
from ..utils import DbtMetadataExtractor

class BotHandler:
    def __init__(self):
        self.config = ConfigHandler()
        self.dbt_extractor = DbtMetadataExtractor('models')
        self._initialize_platforms()

    def _initialize_platforms(self):
        """Initialize enabled messaging platforms."""
        self.platforms = {}
        
        # Initialize Slack if configured
        if self.config.validate_config('slack'):
            slack_config = self.config.get_platform_config('slack')
            self.platforms['slack'] = SlackApp(
                token=slack_config['token'],
                signing_secret=slack_config['signing_secret']
            )

        # Initialize Teams if configured
        if self.config.validate_config('teams'):
            teams_config = self.config.get_platform_config('teams')
            self.platforms['teams'] = connectors.ConnectorClient(
                teams_config['app_id'],
                teams_config['app_password']
            )

        # Initialize Google Chat if configured
        if self.config.validate_config('gchat'):
            gchat_config = self.config.get_platform_config('gchat')
            self.platforms['gchat'] = pubsub_v1.PublisherClient()

    async def handle_message(self, platform: str, message: Dict[str, Any]):
        """Handle incoming messages from any platform."""
        if platform not in self.platforms:
            raise ValueError(f"Platform {platform} not configured")

        # Extract query from message (platform-specific handling)
        query = self._extract_query(platform, message)
        
        # Process the query using dbt metadata
        response = await self._process_query(query)
        
        # Send response (platform-specific handling)
        await self._send_response(platform, message, response)

    def _extract_query(self, platform: str, message: Dict[str, Any]) -> str:
        """Extract query text from platform-specific message format."""
        if platform == 'slack':
            return message.get('text', '')
        elif platform == 'teams':
            return message.get('text', '')
        elif platform == 'gchat':
            return message.get('message', {}).get('text', '')
        return ''

    async def _process_query(self, query: str) -> str:
        """Process the query using dbt metadata."""
        try:
            # Use the dbt metadata extractor to process the query
            metadata = self.dbt_extractor.extract_model_metadata()
            
            # Simple keyword-based response for now
            # This can be enhanced with NLP/ML processing
            if 'model' in query.lower():
                return self._get_model_info(query, metadata)
            elif 'column' in query.lower():
                return self._get_column_info(query, metadata)
            elif 'relationship' in query.lower():
                return self._get_relationship_info(query, metadata)
            else:
                return "I can help you with information about dbt models, columns, and their relationships. What would you like to know?"
        
        except Exception as e:
            return f"Error processing query: {str(e)}"

    async def _send_response(self, platform: str, original_message: Dict[str, Any], response: str):
        """Send response using platform-specific methods."""
        if platform == 'slack':
            await self.platforms['slack'].client.chat_postMessage(
                channel=original_message['channel'],
                text=response
            )
        elif platform == 'teams':
            await self.platforms['teams'].messages.create(
                original_message['channel'],
                content=response
            )
        elif platform == 'gchat':
            topic_path = f"projects/{self.config.gchat_config['project_id']}/topics/{original_message['topic']}"
            self.platforms['gchat'].publish(topic_path, response.encode('utf-8'))

    def _get_model_info(self, query: str, metadata: Dict) -> str:
        """Extract and format model information."""
        # Implementation for model info extraction
        pass

    def _get_column_info(self, query: str, metadata: Dict) -> str:
        """Extract and format column information."""
        # Implementation for column info extraction
        pass

    def _get_relationship_info(self, query: str, metadata: Dict) -> str:
        """Extract and format relationship information."""
        # Implementation for relationship info extraction
        pass 