pyproject.toml
[tool.poetry]
name = "dbt-search-agent"
version = "0.1.0"
description = ""
authors = ["Matteo Gevi <matteogevi97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.39.0"
requests = "^2.32.3"
pandas = "^2.2.3"
slack-sdk = "^3.33.3"
flask = "^3.0.3"
dbt-core = "^1.8.8"
dbt-postgres = "^1.8.2"
ngrok = "^1.4.0"
pyyaml = "^6.0.2"
rasa = "^3.6.15"
rasa-sdk = "^3.6.2"
# Channel connectors
rasa-x = "^1.1.0"  # Optional: for better model management
python-telegram-bot = "^20.8"  # For Telegram
slack-bolt = "^1.18.1"  # For Slack
microsoftteams = "^0.3.0"  # For Microsoft Teams
google-cloud-pubsub = "^2.19.0"  # For Google Chat

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
